backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.12
pydantic[email]==2.9.2
pydantic-settings==2.5.2
slowapi==0.1.9
python-dotenv==1.0.1
